stream: treat compression web stream format per its WebIDL definition
