+++ lib/internal/crypto/webidl.js
@@ -19,19 +19,21 @@ const {
   MathTrunc,
   Number,
   NumberIsFinite,
-  ObjectAssign,
   ObjectPrototypeIsPrototypeOf,
   SafeArrayIterator,
-  SafeSet,
   String,
   SymbolIterator,
   TypedArrayPrototypeGetBuffer,
   TypedArrayPrototypeGetSymbolToStringTag,
-  TypeError,
   globalThis: {
     SharedArrayBuffer,
   },
 } = primordials;
+
+const {
+  makeException,
+  createEnumConverter,
+} = require('internal/webidl');
 
 const {
   kEmptyObject,
@@ -39,23 +41,6 @@ const {
 } = require('internal/util');
 const { CryptoKey } = require('internal/crypto/webcrypto');
 const { getDataViewOrTypedArrayBuffer } = require('internal/crypto/util');
-
-function codedTypeError(message, errorProperties = kEmptyObject) {
-  // eslint-disable-next-line no-restricted-syntax
-  const err = new TypeError(message);
-  ObjectAssign(err, errorProperties);
-  return err;
-}
-
-function makeException(message, opts = kEmptyObject) {
-  const prefix = opts.prefix ? opts.prefix + ': ' : '';
-  const context = opts.context?.length === 0 ?
-    '' : (opts.context ?? 'Value') + ' ';
-  return codedTypeError(
-    `${prefix}${context}${message}`,
-    { code: opts.code || 'ERR_INVALID_ARG_TYPE' },
-  );
-}
 
 // https://tc39.es/ecma262/#sec-tonumber
 function toNumber(value, opts = kEmptyObject) {
@@ -305,22 +290,6 @@ function createDictionaryConverter(name, dictionaries) {
     }
 
     return idlDict;
-  };
-}
-
-function createEnumConverter(name, values) {
-  const E = new SafeSet(values);
-
-  return function(V, opts = kEmptyObject) {
-    const S = String(V);
-
-    if (!E.has(S)) {
-      throw makeException(
-        `value '${S}' is not a valid enum value of type ${name}.`,
-        { __proto__: null, ...opts, code: 'ERR_INVALID_ARG_VALUE' });
-    }
-
-    return S;
   };
 }
