+++ lib/internal/webstreams/compression.js
@@ -5,10 +5,6 @@ const {
 } = primordials;
 
 const {
-  codes: { ERR_INVALID_ARG_VALUE },
-} = require('internal/errors');
-
-const {
   newReadableWritablePairFromDuplex,
 } = require('internal/webstreams/adapters');
 
@@ -19,11 +15,19 @@ const {
   kEnumerableProperty,
 } = require('internal/util');
 
+const { createEnumConverter } = require('internal/webidl');
+
 let zlib;
 function lazyZlib() {
   zlib ??= require('zlib');
   return zlib;
 }
+
+const formatConverter = createEnumConverter('CompressionFormat', [
+  'deflate',
+  'deflate-raw',
+  'gzip',
+]);
 
 /**
  * @typedef {import('./readablestream').ReadableStream} ReadableStream
@@ -38,6 +42,10 @@ class CompressionStream {
    * @param {'deflate'|'deflate-raw'|'gzip'} format
    */
   constructor(format) {
+    format = formatConverter(format, {
+      prefix: "Failed to construct 'CompressionStream'",
+      context: '1st argument',
+    });
     switch (format) {
       case 'deflate':
         this.#handle = lazyZlib().createDeflate();
@@ -48,8 +56,6 @@ class CompressionStream {
       case 'gzip':
         this.#handle = lazyZlib().createGzip();
         break;
-      default:
-        throw new ERR_INVALID_ARG_VALUE('format', format);
     }
     this.#transform = newReadableWritablePairFromDuplex(this.#handle);
   }
@@ -86,6 +92,10 @@ class DecompressionStream {
    * @param {'deflate'|'deflate-raw'|'gzip'} format
    */
   constructor(format) {
+    format = formatConverter(format, {
+      prefix: "Failed to construct 'DecompressionStream'",
+      context: '1st argument',
+    });
     switch (format) {
       case 'deflate':
         this.#handle = lazyZlib().createInflate();
@@ -96,8 +106,6 @@ class DecompressionStream {
       case 'gzip':
         this.#handle = lazyZlib().createGunzip();
         break;
-      default:
-        throw new ERR_INVALID_ARG_VALUE('format', format);
     }
     this.#transform = newReadableWritablePairFromDuplex(this.#handle);
   }
