+++ test/wpt/status/compression.json
@@ -2,25 +2,11 @@
   "compression-bad-chunks.tentative.any.js": {
     "skip": "Execution \"hangs\", ArrayBuffer and TypedArray is not accepted and throws, instead of rejects during writer.write"
   },
-  "compression-constructor-error.tentative.any.js": {
-    "fail": {
-      "expected": [
-        "non-string input should cause the constructor to throw"
-      ]
-    }
-  },
   "decompression-bad-chunks.tentative.any.js": {
     "skip": "Execution \"hangs\", ArrayBuffer and TypedArray is not accepted and throws, instead of rejects during writer.write"
   },
   "decompression-buffersource.tentative.any.js": {
     "skip": "ArrayBuffer and TypedArray is not accepted and throws, instead of rejects during writer.write"
-  },
-  "decompression-constructor-error.tentative.any.js": {
-    "fail": {
-      "expected": [
-        "non-string input should cause the constructor to throw"
-      ]
-    }
   },
   "compression-with-detach.tentative.window.js": {
     "requires": ["crypto"]
